Refactor tests into common package to be included in other packages
