+++ packages/api-electron-demo/package.json
@@ -7,15 +7,16 @@
     "serve": "live-server --port=5000 --no-browser src/",
     "launch": "ssf-electron ./src/app.json",
     "start": "npm-run-all --parallel launch serve",
-    "postinstall": "rimraf src && copyfiles -f \"node_modules/containerjs-api-specification/src/**\" src",
+    "postinstall": "rimraf src && rimraf test && copyfiles -f \"node_modules/containerjs-api-specification/src/**\" src && copyfiles -f \"node_modules/containerjs-api-tests/test/**\" test",
     "test": "echo \"No test specified\" && exit 0",
-    "mocha": "mocha test/init.js"
+    "mocha": "mocha test/*.js"
   },
   "author": "",
   "license": "Apache-2.0",
   "dependencies": {
     "containerjs-api-specification": "0.0.1",
     "containerjs-api-electron": "0.0.1",
+    "containerjs-api-tests": "0.0.1",
     "chai": "^3.5.0",
     "chai-as-promised": "^6.0.0",
     "electron": "^1.6.2",
