+++ packages/api-openfin-demo/package.json
@@ -7,9 +7,9 @@
     "serve": "live-server --port=5000 --no-browser src/",
     "launch": "openfin --launch --config ./src/app.json",
     "start": "npm-run-all --parallel serve launch",
-    "postinstall": "rimraf src && copyfiles -f \"node_modules/containerjs-api-specification/src/**\" src && copyfiles -f \"node_modules/containerjs-api-openfin/dist/**\" src",
+    "postinstall": "rimraf src && rimraf test && copyfiles -f \"node_modules/containerjs-api-specification/src/**\" src && copyfiles -f \"node_modules/containerjs-api-openfin/dist/**\" src && copyfiles -f \"node_modules/containerjs-api-tests/test/**\" test",
     "test": "echo \"No test specified\" && exit 0",
-    "mocha": "mocha test/init.js"
+    "mocha": "mocha test/*.js"
   },
   "author": "",
   "license": "Apache-2.0",
@@ -20,6 +20,7 @@
     "openfin-cli": "^1.1.5",
     "containerjs-api-specification": "0.0.1",
     "containerjs-api-openfin": "0.0.1",
+    "containerjs-api-tests": "0.0.1",
     "selenium-webdriver": "^2.44.0",
     "should": "^4.1.0",
     "wd": "^0.3.10",
