+++ packages/api-tests/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "containerjs-api-tests",
+  "version": "0.0.1",
+  "license": "Apache-2.0"
+}
